refactor(vsock): split virtio device internals
